Regenerate Docker conda environment (#138)
* Regenerate conda lockfile

Refresh the conda environment used by the Docker image to fetch the
latest patches in preparation for the upcoming DISP-S1 release.

* Add a basic CI workflow to test Docker builds

Adds a simple GHA workflow that runs on pull requests, pushes to the
default branch and release branches, and tags. It tries to build a
Docker image from the Dockerfile and runs a simple smoke test using the
resulting image.

* Remove an extra line in conda lockfile

`create-lockfile.sh` shouldn't print anything to the console besides the
contents of the lockfile. Doing so breaks its intended usage pattern --
you create the lockfile by piping its output to a text file.